Move bulk status dialog to bulk folder
